FerMod

+++ src/main/java/com/ruralhousejsf/converter/RuralHouseConverter.java
@@ -6,14 +6,14 @@ import javax.faces.convert.FacesConverter;
 
 import org.omnifaces.converter.SelectItemsConverter;
 
-import domain.RuralHouse;
+import com.ruralhousejsf.domain.RuralHouse;
 
 @FacesConverter("ruralHouseConverter")
 public class RuralHouseConverter extends SelectItemsConverter  {
 
 	@Override
-	public String getAsString(FacesContext context, UIComponent component, Object value) {		
-		Integer id = (value instanceof RuralHouse) ? ((RuralHouse) value).getId() : null;
+	public String getAsString(FacesContext context, UIComponent component, Object value) {	
+		Long id = (value instanceof RuralHouse) ? ((RuralHouse) value).getId() : null;
 		return (id != null) ? String.valueOf(id) : null;
 	}
